config: extract zone compilation to compiler_security_zones.go
Part B of #4121: compiler_security.go had grown into a ~2357-line
grab-bag mixing zones, policies, screen/IDS, address-book, log, flow,
and ALG compilation plus a dozen commit-time validators. Split it into
focused per-concern modules for maintainability. This is PURE CODE
MOTION — no logic change, no behavior change. Part A (the compilePolicy
match-value SSOT consolidation) shipped separately in #4133.

This commit moves the zone-compilation cluster verbatim:

  - parseHostInboundNode  (shared SSOT for zone-level and per-interface
    host-inbound-traffic parsing, #3362)
  - compileZones

Function bodies are byte-identical to the origin/master original; only
the enclosing file and its package/import boilerplate changed. Both
functions stay in package config, so no import changes are needed for
intra-package references (parseAddressBookEntries, namedInstances,
firewallMatchValues, etc. resolve across the same package).

Validation: gofmt -l clean, go vet ./pkg/config/, go build ./...,
go test ./pkg/config/... green on the fully-split tree.
